# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Poachy API Documentation",
        "description": "Multi-tenant POS and eCommerce Marketplace Platform API",
        "contact": {
            "name": "Poachy Support",
            "email": "support@poachy.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://poachy.com/license"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/api/v1/central/tenants/{tenant_id}": {
            "get": {
                "tags": [
                    "Tenant Management"
                ],
                "summary": "Get Tenant Details",
                "description": "Get detailed information about a specific tenant including domains and business details.",
                "operationId": "getTenant",
                "parameters": [
                    {
                        "name": "tenant_id",
                        "in": "path",
                        "description": "Tenant UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant details retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tenant retrieved successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/TenantResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Resource not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Tenant Management"
                ],
                "summary": "Delete Tenant",
                "description": "Delete a tenant and all associated data including database. This action is irreversible.",
                "operationId": "deleteTenant",
                "parameters": [
                    {
                        "name": "tenant_id",
                        "in": "path",
                        "description": "Tenant UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Tenant deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Tenant deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/central/tenants/search": {
            "get": {
                "tags": [
                    "Tenant Management"
                ],
                "summary": "Search Tenants",
                "description": "Search tenants by domain name or business name.",
                "operationId": "searchTenants",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Search query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "merchant"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Search results retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Search results retrieved successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/central/tenants/{tenant_id}/domains": {
            "post": {
                "tags": [
                    "Tenant Management"
                ],
                "summary": "Add Domain to Tenant",
                "description": "Add an additional domain to an existing tenant.",
                "operationId": "addDomainToTenant",
                "parameters": [
                    {
                        "name": "tenant_id",
                        "in": "path",
                        "description": "Tenant UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "domain"
                                ],
                                "properties": {
                                    "domain": {
                                        "type": "string",
                                        "example": "merchant1-new.poachy.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Domain added successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Domain added successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/DomainResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Tenant not found"
                    },
                    "422": {
                        "description": "Validation error - Domain already exists"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DomainResource": {
                "title": "Domain Resource",
                "description": "Domain resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "domain": {
                        "type": "string",
                        "example": "merchant.poachy.com"
                    },
                    "tenant_id": {
                        "type": "string",
                        "format": "uuid",
                        "example": "9d8e4f2a-1b3c-4d5e-6f7a-8b9c0d1e2f3a"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Laravel Sanctum token authentication. Enter your token in the format: Bearer {token}",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}